+++ id3tag/src/default_values.rs
@@ -138,10 +138,7 @@ impl DefaultValues {
                     .unwrap_or("~/.config/id3tag/config.toml"),
             )
             .to_string();
-            log::debug!("build_config::config_filename = {}", config_filename);
             config = Self::load_config(&config_filename)?;
-        } else {
-            log::debug!("No config file specified.");
         }
 
         // Collate config file flags and CLI flags and output the right config
@@ -155,7 +152,6 @@ impl DefaultValues {
         config.check_for_picture_max_size(cli_args);
         config.check_for_picture_front_candidates(cli_args);
         config.check_for_picture_back_candidates(cli_args);
-        log::debug!("Working config: {:?}", &config);
 
         Ok(config)
     }
@@ -163,20 +159,14 @@ impl DefaultValues {
     /// Loads the config from the supplied TOML file.
     pub fn load_config(filename: &str) -> Result<Self, Box<dyn Error>> {
         let mut config_toml = String::new();
-        log::debug!("load_config::Loading {filename}");
 
         let mut file = File::open(filename)
             .map_err(|err| format!("Config file {filename} not found. Error: {err}"))?;
 
-        log::debug!("load_config::file = {file:?}");
-
         let bytes = file.read_to_string(&mut config_toml)?;
-        if bytes > 0 {
-            log::debug!("Bytes read: {bytes}");
-        } else {
+        if bytes == 0 {
             return Err(format!("Unable to read the contents of {filename}").into());
         }
-        log::debug!("load_config::config_toml = {config_toml}");
 
         let mut config = match toml::from_str(&config_toml) {
             Ok(config) => config,
@@ -190,8 +180,6 @@ impl DefaultValues {
             }
         };
 
-        log::debug!("load_config::config = {config:?}");
-
         // Check if the picture_search_folders contain "." and "..". Add them if not.
         let mut psf = config.picture_search_folders.clone().unwrap_or_default();
         if !psf.contains(&'.'.to_string()) {
@@ -201,8 +189,6 @@ impl DefaultValues {
             psf.push("..".to_string());
         }
         config.picture_search_folders = Some(psf);
-
-        log::debug!("load_config::returning Ok({config:?})");
 
         Ok(config)
     }