+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 
 - Remove unused debugs
 - Changelog
+- Changelog
 
 ## [0.14.2] - 2023-02-20